Setup a new cluster for dev, introduce script/login.sh

+++ script/create_instance.sh
@@ -1,13 +1,12 @@
 #!/bin/bash
 
 set -x -v
+set -e
 
-source ./script/env_source.sh
+TARGET_ENV="${TARGET_ENV:-dev}"
+source ./.env.$TARGET_ENV
 
-#
-# Login account
-az login
-az account set -s "${SUBSCRIPTION_ID}"
+./script/login.sh || echo "warning: login issues but will continue"
 
 # # Create the Azure AD application
 # serverApplicationId="$(az ad app list| jq '[.[]| { name: .displayName, id: .appId}]' | jq -r '.[]|select("'${CLUSTER_NAME}Server'" == .name)|.id')"
@@ -94,14 +93,21 @@ tenantId=$(az account show --query tenantId -o tsv)
 
 # From https://github.com/Azure/azure-cli/issues/9585#issuecomment-502542000
 credId="$(az ad app list| jq '[.[]| { name: .displayName, id: .appId}]' | jq -r '.[]|select("'${CLUSTER_NAME}AKS-sp'" == .name)|.id')"
-if [ ! -z "${credId}" ]; then
-  az ad sp delete --id "${credId}"
-fi
-az ad sp create-for-rbac \
-    --name "${CLUSTER_NAME}AKS-sp" \
-    --skip-assignment > "${AD_SP_CREDS_JSON}"
-credId="$(az ad app list| jq '[.[]| { name: .displayName, id: .appId}]' | jq -r '.[]|select("'${CLUSTER_NAME}AKS-sp'" == .name)|.id')"
 
+# if the cluster exist, don't create a new id, just re-use the existing one
+cluster_id="$(az aks list| jq -r '.[] | select("'${CLUSTER_NAME}'" == .name) | .id ')"
+if [ -z "${cluster_id}" ]; then
+    if [ ! -z "${credId}" ]; then
+        az ad sp delete --id "${credId}"
+    fi
+    az ad sp create-for-rbac \
+        --name "${CLUSTER_NAME}AKS-sp" \
+        --skip-assignment > "${AD_SP_CREDS_JSON}"
+    credId="$(az ad app list| jq '[.[]| { name: .displayName, id: .appId}]' | jq -r '.[]|select("'${CLUSTER_NAME}AKS-sp'" == .name)|.id')"
+fi
+
+# TODO: these are not being used but might be useful for us later when working with AAD managed clusters
+# we're leaving it out for now.
 # serverApplicationId="$(az ad app list| jq '[.[]| { name: .displayName, id: .appId}]' | jq -r '.[]|select("'${CLUSTER_NAME}Server'" == .name)|.id')"
 # if [ -z "${serverApplicationId}" ]; then
 # serverApplicationId=$(az ad app create \
@@ -131,15 +137,16 @@ az aks create \
     --node-count 1 \
     --min-count 1 \
     --max-count $NODE_COUNT \
-    --load-balancer-sku basic \
+    --load-balancer-sku $SKU_NAME \
     --generate-ssh-keys \
     --network-plugin azure \
     --service-principal "${credId}" \
-    --client-secret "$(cat ~/.azcli_creds.secrets.json| jq -r '.password')" \
-    --enable-vmss \
+    --client-secret "$(cat "${AD_SP_CREDS_JSON}" | jq -r '.password')" \
+    --vm-set-type VirtualMachineScaleSets \
     --enable-cluster-autoscaler \
     --subscription "${SUBSCRIPTION_ID}"
 
+# TODO: these are not being used but might be useful for us later when working with AAD managed clusters
     # --node-vm-size "${VM_SIZE}" \
     # --node-count "${NODE_COUNT}" \
     # --vm-set-type AvailabilitySet \
@@ -157,7 +164,7 @@ az aks create \
     # --client-secret null \
     # --subscription "${SUBSCRIPTION_ID}"
 
-az aks show -g $RESOURCE_GROUP -n $CLUSTER_NAME \
-                                --query "identity" \
-                                --subscription $SUBSCRIPTION_ID
+echo "AKS version ==> $(az aks show -g $RESOURCE_GROUP -n $CLUSTER_NAME \
+                                --query "currentKubernetesVersion" \
+                                --subscription $SUBSCRIPTION_ID)"
